data: Add simple URI data fetcher
Instead of downloading and checking-in all the data for the project, use
a script that can download and unzip all our data sources for this
exercise.

Unzipping is integrated for mere convenience, given our very specific
data sources.

The list of URIs to download is handled as a simple python tuple of
tuples in fetch_data_uris.py.