bc/views: Add /display_data/ endpoint
Implement a bare bones /display_data/ endpoint that returns JSON
responses to different GET parameters.

Supported GET parameters are: country, indicator, year. Parameters can
be combined for more specific sets of results.

A followup to this view should be able to handle ranges and lists in the
GET parameters so you can query for "x, y countries since a to b years
and m and n indicators". This should become more obvious and be inspired
by the needs of the frontend.

For the purposes of this demo, the default behavior is to return the
contents of the whole database.