frontvue: Import our proof of concept frontend
This also integrates the `yarn` tooling into the static path of
backcountry, so that both front and backend can be served through
Django's runserver.

Note that for development `yarn serve` continues to be the easiest
route, coupled with `manage.py runserver` for the backend.

We won't be checking in the output of `yarn build` except for milestone
versions/git tags, this only to keep things simple and easy for
development and testing.