fv: Add ChunkyTable component
Cleanup TrendyChart to be smarter and more agnostic of how it's being
used, and add ChunkyTable to show datasets in chunks of 10 (decades).

This makes our data presentation situation "good enough" to work on
PATCH requests and move into cosmetic and interactive improvements.

Ideally we should be able to reuse these two components with formatting
functions or similar tricks, so that each table and chart present in
nicely formatted numbers, following the user locale.