fv: Bubble-up update signal from EditableCell
We now raise EditableCell::update:value up to
ChunkyTable::update:datasets and sync it up manually to TrendyChart in
CountryStats.

Ideally we would get the updates from the server and ChunkyTable and
TrendyChart would be updated as soon as the backend gets PATCH'd.
However that would complicate things a bit more so we are going manual
for now.